chore: 클래스명 정리
