+++ src/main/java/subway/path/application/PathStrategy.java
@@ -8,11 +8,7 @@ import subway.station.domain.Station;
 
 import java.util.List;
 
-public interface PathStrategy {  // TODO
-//    Double getWeightOfPath(WeightedMultigraph<Station, DefaultWeightedEdge> graph, Station sourceStation, Station targetStation);
-
-    Long getTotalWeightInPath(List<Station> stationsInPath, List<Section> sections);
-
+public interface PathStrategy {
     void setEdgeWeight(WeightedMultigraph<Station, DefaultWeightedEdge> graph, Section section, DefaultWeightedEdge edge);
 
     PathRetrieveResponse findPath(WeightedMultigraph<Station, DefaultWeightedEdge> graph, List<Section> sections, List<Station> stationsInPath, Station sourceStation, Station targetStation);
