+++ src/test/java/subway/unit/path/ShortestDistancePathTest.java
@@ -7,8 +7,8 @@ import subway.acceptance.station.StationFixture;
 import subway.exception.SubwayBadRequestException;
 import subway.line.domain.Line;
 import subway.line.domain.Section;
-import subway.path.application.MinimumTimePathStrategy;
 import subway.path.application.PathFinder;
+import subway.path.application.ShortestDistancePathFinder;
 import subway.path.application.dto.PathRetrieveResponse;
 import subway.station.application.dto.StationResponse;
 import subway.station.domain.Station;
@@ -20,10 +20,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static subway.acceptance.station.StationFixture.getStation;
 
-@DisplayName("MinimumTimePathFinder 단위 테스트")
-public class MinimumTimePathStrategyTest {
-    private PathFinder minimumTimePathFinder;
-
+@DisplayName("ShortestDistancePathFinder 단위 테스트")
+public class ShortestDistancePathTest {
+    private PathFinder shortestDistancePathFinder;
     private Line 이호선;
     private Line 삼호선;
     private Line 신분당선;
@@ -46,8 +45,8 @@ public class MinimumTimePathStrategyTest {
 
     @BeforeEach
     void beforeEach() {
-        MinimumTimePathStrategy minimumTimePathStrategy = new MinimumTimePathStrategy();
-        minimumTimePathFinder = new PathFinder(minimumTimePathStrategy);
+        ShortestDistancePathFinder shortestDistancePathStrategy = new ShortestDistancePathFinder();
+        shortestDistancePathFinder = new PathFinder(shortestDistancePathStrategy);
 
         StationFixture.기본_역_생성();
 
@@ -72,25 +71,26 @@ public class MinimumTimePathStrategyTest {
 
     }
 
+
     /**
      * Given 구간이 있을 때
      * When 경로 조회를 하면
      * Then 경로 내 역이 목록으로 반환된다
-     * Then 경로의 총 시간이 반환된다
+     * Then 경로의 총 길이가 반환된다
      */
-    @DisplayName("최소시간 경로 조회")
+    @DisplayName("최단거리 경로 조회")
     @Test
-    void getMinimumTimePath() {
+    void getShortestDistancePath() {
         // when
-        PathRetrieveResponse shortestPath = minimumTimePathFinder.findPath(구간목록, getStation("교대역"), getStation("양재역"));
+        PathRetrieveResponse shortestPath = shortestDistancePathFinder.findPath(구간목록, getStation("교대역"), getStation("양재역"));
 
         // then
         assertThat(shortestPath.getStations())
                 .extracting(StationResponse::getName)
-                .containsExactlyInAnyOrder("교대역", "강남역", "양재역");
+                .containsExactlyInAnyOrder("교대역", "남부터미널역", "양재역");
 
         // then
-        assertThat(shortestPath.getDuration()).isEqualTo(11L);
+        assertThat(shortestPath.getDistance()).isEqualTo(5L);
     }
 
     /**
@@ -98,11 +98,11 @@ public class MinimumTimePathStrategyTest {
      * When 동일한 구간을 조회하면
      * Then 조회 되지 않는다.
      */
-    @DisplayName("최소시간 경로 조회 : 요청 구간 동일")
+    @DisplayName("최단거리 경로 조회 : 요청 구간 동일")
     @Test
-    void getMinimumTimePathWithSameOrigin() {
+    void getShortestDistancePathWithSameOrigin() {
         // when/then
-        assertThatThrownBy(() -> minimumTimePathFinder.findPath(구간목록, getStation("교대역"), getStation("교대역")))
+        assertThatThrownBy(() -> shortestDistancePathFinder.findPath(구간목록, getStation("교대역"), getStation("교대역")))
                 .isInstanceOf(SubwayBadRequestException.class);
     }
 
@@ -111,11 +111,11 @@ public class MinimumTimePathStrategyTest {
      * When 연결되지 않은 구간을 조회하면
      * Then 조회 되지 않는다.
      */
-    @DisplayName("최소시간 경로 조회 : 연결되지 않은 구간")
+    @DisplayName("최단거리 경로 조회 : 연결되지 않은 구간")
     @Test
-    void getMinimumTimePathNotConnectedSection() {
+    void getShortestDistancePathNotConnectedSection() {
         // when/then
-        assertThatThrownBy(() -> minimumTimePathFinder.findPath(구간목록, getStation("교대역"), getStation("왕십리역")))
+        assertThatThrownBy(() -> shortestDistancePathFinder.findPath(구간목록, getStation("교대역"), getStation("왕십리역")))
                 .isInstanceOf(SubwayBadRequestException.class);
     }
 
@@ -124,13 +124,12 @@ public class MinimumTimePathStrategyTest {
      * When 존재하지 않는 역을 조회하면
      * Then 조회 되지 않는다.
      */
-    @DisplayName("최소시간 경로 조회 : 존재하지 않는 역")
+    @DisplayName("최단거리 경로 조회 : 존재하지 않는 역")
     @Test
-    void getMinimumTimePathNotExistStation() {
+    void getShortestDistancePathNotExistStation() {
         // when/then
-        assertThatThrownBy(() -> minimumTimePathFinder.findPath(구간목록, new Station(99L, "그런역"), new Station(98L, "저런역")))
+        assertThatThrownBy(() -> shortestDistancePathFinder.findPath(구간목록, new Station(99L, "그런역"), new Station(98L, "저런역")))
                 .isInstanceOf(SubwayBadRequestException.class);
     }
-
 
 }
